requirements.txt
# 小说整理系统GUI - 依赖包列表

# 核心依赖
PyYAML>=6.0.0          # YAML配置文件处理
chardet>=5.0.0         # 文件编码检测
colorlog>=6.0.0        # 彩色日志输出

# GUI相关 (Tkinter是Python标准库，无需安装)
# tkinter (标准库)

# 可选依赖
# Pillow>=9.0.0        # 图像处理 (如果需要图标支持)
# psutil>=5.8.0        # 系统监控 (如果需要性能监控)

# 开发依赖 (可选)
# pytest>=6.2.0       # 测试框架
# black>=21.0.0        # 代码格式化
# flake8>=3.9.0        # 代码检查
